Move generic components to common
