+++ client/src/components/common/input/RadioButtonSet.jsx
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react'
 import _ from 'lodash'
 
-import FlexRow from '../../common/ui/FlexRow'
+import FlexRow from '../ui/FlexRow'
 import {FilterColors} from '../../../style/defaultStyles'
 
 const styleRadioTab = {
